+++ CHANGELOG.md
@@ -11,9 +11,10 @@
  - Emit error but set default when feed rate not set and cutting move is made.
  - Allow dot in GCode variable reference names.
  - Warn about rendering empty workpiece.
+ - Move all axes on "G28/G30 <axes>" as specified in RS274/NGC section 3.5.8.
 
 ## v1.2.0:
- - Improved Buildbotics CNC controller connectivity.
+- Improved Buildbotics CNC controller connectivity.
  - Revamped donation dialog.
  - Show donation dialog once at start with each upgrade.
  - S-Curve path planner.